utils/requirements.txt
pillow      # For image handling in seetex.py
pypng       # For PNGs in texy.py
